WS30 - Improve line feature loading
- Mark tiles that require re-loading after .STG loading, removing a race
condition
- Only re-build line features rather than the entire tile

+++ simgear/scene/tgdb/LineFeatureBin.cxx
@@ -82,7 +82,7 @@ LineFeatureBin::LineFeatureBin(const SGPath& absoluteFileName, const std::string
         }
 
         while (true) {
-            float lon = 0.0f, lat=0.0f;
+            double lon = 0.0f, lat=0.0f;
             in >> lon >> lat;
 
             if (in.bad() || in.fail()) {
